# docs/api/coordinator.openapi.json
{
  "components": {
    "schemas": {
      "StartAnalysisRequest": {
        "properties": {
          "configId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "configId"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "Coordinator",
    "version": "0.0.1"
  },
  "openapi": "3.0.0",
  "paths": {
    "/analysis/start": {
      "post": {
        "operationId": "AnalysisController_start",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Analysis"
        ]
      }
    },
    "/analysis/{runId}/retry/{cluster}": {
      "post": {
        "operationId": "AnalysisController_retry",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "cluster",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Analysis"
        ]
      }
    },
    "/analysis/{runId}/stream": {
      "get": {
        "operationId": "AnalysisController_stream",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analysis"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/simulation/statuses": {
      "get": {
        "operationId": "SimulationController_getStatuses",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Simulation"
        ]
      }
    },
    "/simulation/{cluster}/down": {
      "post": {
        "operationId": "SimulationController_down",
        "parameters": [
          {
            "in": "path",
            "name": "cluster",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Simulation"
        ]
      }
    },
    "/simulation/{cluster}/up": {
      "post": {
        "operationId": "SimulationController_up",
        "parameters": [
          {
            "in": "path",
            "name": "cluster",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Simulation"
        ]
      }
    }
  },
  "servers": [],
  "tags": []
}